* commit '7f1e9212655971570662c4bb93f85e456cc2cebf':
  Further specifies src folders.
  Fixes some Moduletests
  Removes .vscode folder
  Adds Response on Event and LinkControl Tests.
  Removes reference to ResponseCode.h
  Adds Coverage Report.
  Deactivates Windows Port Test
  Removes second checkout.
  Now uses correct Branch.
  Adds cleanup of old workspace.
  Fixes Ceedling Tests.
  Fixes Git Credentials.
  Adds rudimentary Jenkinsfile.
  Adds Documentation and unique name convention.

+++ test/support/TestTime.h
@@ -5,6 +5,6 @@
 
 uint32_t time_getTime(void);
 
-int32_t time_diffTime(uint32_t _this, uint32_t _that);
+int64_t time_diffTime(uint32_t _this, uint32_t _that);
 
 #endif
